+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.6.5](https://github.com/Ash258/Scoop-Core/milestone/5)
 
+- Initial support for manipulation with `arm64` based manifests
 - **scoop-search**:
     - Do not fail when parsing invalid local manifest
     - Support `githubToken` config and `GITHUB_TOKEN` environment variable for Github API calls